feat: port `p2p_libp2p` connection tests to the package
