+++ packages/valory/agents/test_libp2p/tests/test_fault_tolerance.py
@@ -22,14 +22,11 @@
 import pytest
 
 from aea.multiplexer import Empty, Multiplexer
+from aea.test_tools.utils import wait_for_condition
 
+from packages.valory.agents.test_libp2p.tests.base import BaseP2PLibp2pTest
 from packages.valory.connections.p2p_libp2p.check_dependencies import build_node
-
-from tests.common.utils import wait_for_condition
-from tests.test_packages.test_connections.test_p2p_libp2p.base import (
-    BaseP2PLibp2pTest,
-    libp2p_log_on_failure_all,
-)
+from packages.valory.connections.p2p_libp2p.tests.base import libp2p_log_on_failure_all
 
 
 TIMEOUT = 10
