+++ packages/valory/connections/p2p_libp2p/connection.yaml
@@ -54,10 +54,17 @@ fingerprint:
   libp2p_node/protocols/acn/v1_0_0/acn.yaml: bafybeidcmqt4wad37njvr5rzqlx2rlqjiviq6dpepkvbvhf2xwqcnl4jda
   libp2p_node/utils/utils.go: bafybeihbd2br74nm3pupy4qkijm2tnuyl7pbq3uwasidmql5znhbr575k4
   libp2p_node/utils/utils_test.go: bafybeig2kkgqy7miml66w2byomeqjp4czzv45gb4jas76cxildo3lxi4xq
+  tests/__init__.py: bafybeieftcbmxxpe7okvm3ycualpyec6xys4nx5tihcjii7lqxd3w5lx7e
+  tests/base.py: bafybeig74xjeutlhlgh2j3y6aapgtmm54ancom5rfqddwq5wlgky64nkzu
+  tests/test_aea_cli.py: bafybeick26kvlx72qhayo42salsnrjd5b7oiihfc2pouoxd4qalfl25l2i
+  tests/test_build.py: bafybeibym7xuidezyekqukjepfqsandjukzgsavd3bvudnlrc25ot5pm34
+  tests/test_errors.py: bafybeiga42lt5gb372k4qmvxcw4bui3hyg2up425lmapu26hlfrpnk2e5m
+  tests/test_go_code_matching_acn.py: bafybeihyfb5upmo4ifvbngx46h7h52ale5sahuxyx3ssftv7pfv4xodhwa
 fingerprint_ignore_patterns: []
 build_entrypoint: check_dependencies.py
 connections: []
 protocols:
+- fetchai/default:1.0.0:bafybeihzesahyayexkhk26fg7rqnjuqaab3bmcijtjekvskvs4xw6ecyuu
 - valory/acn:1.1.0:bafybeih5pz5re4kycou46xbzgguantxzj46aqtc575pgvej3m6bfg4xpfy
 class_name: P2PLibp2pConnection
 config:
